requirements.txt
# Create env:
# python3 -m venv mpai_venv
# source mpai_venv/bin/activate
# python3 -m pip install -r requirements.txt

geopy=2.2.0
matplotlib-base=3.5.2
networkx=2.8.5
numpy=1.23.1
pandas=1.4.3
plotnine=0.9.0
ruptures=1.1.7
scipy=1.9.0
seaborn=0.11.2
seaborn-base=0.11.2
six=1.16.0
statsmodels=0.13.2